updated mixins to 1.20.1

+++ src/main/java/com/midnightdoggo19/singleplayer/mixin/BreakingSplashMixin.java
@@ -13,5 +13,5 @@ public class BreakingSplashMixin {
      * @reason Breaking a splash.
      */
     @Overwrite
-    public void render (DrawContext context, int screenWidth, TextRenderer textRenderer, float alpha) {}
+    public void render(DrawContext context, int screenWidth, TextRenderer textRenderer, int alpha) {}
 }
